refactor(utils): Move separated methods to lib/utils
